+++ examples/product_showcase/lib/models/latest_dart_derive_showcase.dart
@@ -1,5 +1,5 @@
 import 'package:dust_dart/db.dart'
-    show FromRow, Row, Sqlx, SqlxRename, registerRowMapper;
+    show FromRow, Row, RowDeserializer, Sqlx, SqlxRename;
 import 'package:dust_dart/serde.dart';
 
 part 'latest_dart_derive_showcase.g.dart';